requirements.txt
deap==1.4.3
numpy==2.2.5
PyYAML==6.0.2
